working on new float tests, some errors are presented

+++ src/fast_convolution/clib/libs/convolution.h
@@ -12,10 +12,10 @@ void fast_conv_float(float *ms, const float *ma, const float *mgg, const float *
                      int c_size);
 
 void
-filter1d_slide1d_float(float *feature_out, const int *feature_in, int index, const float *mc, const float *ma,
+filter1d_slide1d_float(float *feature_out, const float *feature_in, int index, const float *mc, const float *ma,
                        const float *mgg, int a_size, int c_size, int fin_size, int fout_size);
 void
-filter2d_slide2d_float(float *feature_out, const int *feature_in, const float *mc, const float *ma, const float *mgg,
+filter2d_slide2d_float(float *feature_out, const float *feature_in, const float *mc, const float *ma, const float *mgg,
                        int a_size, int c_size, int fin_size, int fout_size);
 
 void matrix_mul(int *out, const int *in1, const int *in2, int row1, int col2_row1, int col2);
@@ -27,6 +27,6 @@ void matrix_transpose(int *out, const int *in, int row, int col);
 void hadamart_product(int *out, const int *in1, const int *in2, int row);
 void hadamart_product_float(float *out, const float *in1, const float *in2, int row);
 
-void to_bg(float *mgg, const int *mq, const int *mb, const int *mg, int b_size, int c_size);
+void to_bg(float *mgg, const float *mq, const float *mb, const float *mg, int b_size, int c_size);
 
 #endif //C_CONVOLUTION_H
